% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luc_colors.R
\name{scale_color_luc_c}
\alias{scale_color_luc_c}
\alias{scale_fill_luc_c}
\alias{scale_color_luc_cd}
\alias{scale_fill_luc_cd}
\alias{scale_color_luc_d}
\alias{scale_fill_luc_d}
\title{Loyola Chicago color palettes}
\usage{
scale_color_luc_c(reverse = FALSE, ...)

scale_fill_luc_c(reverse = FALSE, ...)

scale_color_luc_cd(reverse = FALSE, ...)

scale_fill_luc_cd(reverse = FALSE, ...)

scale_color_luc_d(...)

scale_fill_luc_d(...)
}
\arguments{
\item{reverse}{reverse color order}

\item{...}{additional method arguments}
}
\description{
Loyola University Chicago themed \code{ggplot2} color palettes
}
\examples{

if (require(ggplot2)) {
  ggplot(mtcars, aes(hp, mpg, color = mpg)) +
    geom_point(size = 4) +
    scale_color_luc_c(reverse = TRUE)
}
}
\references{
\url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}

\url{https://www.luc.edu/umc/brandstandards/graphicidentity/colorpalette/}
}
